query test

+++ src/Controller/HomeController.php
@@ -1,30 +1,41 @@
 <?php
 namespace App\Controller;
 
+
+use App\Entity\Medicines;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 class HomeController extends AbstractController {
     /**
-     * @Route("/")
+     * @Route("/", name="homepage")
      */
     public function showHome() {
-        return new Response("Bruh");
+        return $this->render('home.html.twig');
     }
     /**
-     * @Route("door/{title}")
+     * @Route("door/{title}", name="commentSection")
      */
-    public function show($title) {
+    public function show($title, EntityManagerInterface $em) {
+        $repository = $em->getRepository(Medicines::class);
+        $medicine = $repository->findOneBy(['id' => 1]);
+
+        if (!$medicine) {
+            throw $this->createNotFoundException(sprintf("no Medicine found"));
+        }
+
         $comments = [
             'why die?',
             'just live',
             'eat sht',
             'I kill problems with knowledge'
         ];
-        return $this->render('article/show.html.twig', [
+        return $this->render('showComments.html.twig', [
             'title' => ucwords(str_replace('-',' ', $title)),
-            'comments' => $comments
+            'comments' => $comments,
+            'medicine' => $medicine
         ]);
     }
     /**
